requirements.txt
autopep8==1.2.4
backports.shutil-get-terminal-size==1.0.0
dateutils==0.6.6
decorator==4.0.10
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pep8==1.7.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.5
ptyprocess==0.5.1
PyEmail==0.0.1
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.11.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
